Participate module started
