+++ static/scripts/user_dashboard.js
@@ -2,6 +2,17 @@ const myevents = document.getElementById('myevents')
 const invitations = document.getElementById('invitations')
 const logoutbtn = document.getElementById('logoutbtn')
 
+function get_cookie_dict(cookie_str) {
+    let cookie_dict = {}
+    let all_cookies = document.cookie.split(';')
+    all_cookies.forEach(element => {
+        let key_value = element.trim().split('=')
+        cookie_dict[key_value[0]] = key_value[1]
+    });
+
+    return cookie_dict
+}
+
 myevents.addEventListener('click', function (event) {
     event.preventDefault();
     console.log('my events called!')
@@ -53,3 +64,40 @@ logoutbtn.addEventListener('click', function (event) {
         console.log(data)
     })
 })
+
+try {
+    const participateBtns = document.getElementsByName('participateBtn')
+
+    const cookie_dict = get_cookie_dict(document.cookie)
+
+    participateBtns.forEach(participateBtn => {
+        participateBtn.addEventListener('click',function (event) {
+            event.preventDefault();
+    
+            if (cookie_dict['access'] == undefined || (('access' in cookie_dict) == false)) {
+                window.location.href = 'http://127.0.0.1:8000/ems/user/login/'
+            }else{
+    
+                data = {
+                    event_id: participateBtn.id,
+                    invite_to: participateBtn.dataset.host,
+                }
+    
+                fetch('http://127.0.0.1:8000/api/invite/',{
+                    method: 'POST',
+                    headers: {
+                        'Authorization': `Bearer ${cookie_dict['access']}`,
+                        'Content-Type': 'application/json',
+                    },
+                    body: JSON.stringify({
+                        data
+                    })
+                })
+            }
+        })
+    });
+
+    
+} catch (error) {
+    
+}
